+++ packages/custom_lint_core/lib/src/lint_codes.dart
@@ -4,7 +4,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:analyzer/error/error.dart';
+import 'package:analyzer/error/error.dart' hide LintCode;
 import 'package:meta/meta.dart';
 
 import '../custom_lint_core.dart';